Fixed a series of implicit type conversion compiler warnings.
Former-commit-id: ab4efef76347f587b0152cec590b5d85f3c504da
Former-commit-id: 53ca4dcdf9d5c0d67a927f6f3580136537f83d04

+++ ADCSAlgorithms/attDetermination/InertialUKF/attitude_ukf.cpp
@@ -165,7 +165,7 @@ void STInertialUKF::UpdateState(uint64_t callTime)
     MatrixOperations inputMatrix;
     vehicleConfigData localConfig;
     
-    ReadMessage(stInputID, &ClockTime, &ReadSize, sizeof(STOutputData), &stMeas, moduleID);
+    ReadMessage((uint32_t)stInputID, &ClockTime, &ReadSize, sizeof(STOutputData), &stMeas, moduleID);
     ReadMessage(inputSpeedsID, &ClockTime, &ReadSize, sizeof(RWSpeedData), &currentSpeeds, moduleID);
     ReadMessage(inputVehicleConfigDataID, &ClockTime, &ReadSize,
         sizeof(vehicleConfigData), &localConfig, moduleID);
